Normalize collapsed nav flyouts
The collapsed navigation bar has two different visual representations
for items today. For navigation items with children there are flyouts,
if there are no children a popover is used instead. The visual
difference is distracting.

Now a flyout is always used. If there are no children, it exists without
children. The flyout header text is now always a link in addition to the
existing icon link. For items with children, this was redundant with the
first child item. Now the 'Overview' item is contributed to the sub nav
automatically when the nav bar is expanded.

Signed-off-by: Scott Andrews <andrewssc@vmware.com>